requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.30.3
packaging==24.0
pefile==2023.2.7
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.32.2
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
